pyproject.toml
[tool.poetry]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.10.0,<3.11"
python3-nmap = "1.6.0"
pip-install = "1.3.5"
socket-plotter = "0.1.4"
pyshark = "0.6"
requests = "2.31.0"
scapy = "2.5.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"